Refactored Dictionary and TotaltCompensation
Added values for additional dependents under and over 18 per percentage rate to the dictionary. Getter methods for use in calculation class added to the dictionary.  Less spaghetti code now and things looks a little cleaner and more organized. My first app was hot trash I am seeing. This looks a lot better and should be easier to maintain and to update along the way.