BANKERS ROUNDING?! ARGHHHH
it took me way to long to figure out that C# Round() built in method uses bankers rounding and was rounding my VA percentages DOWN to the nearest even number.  Why... ON EARTH would that be the default?! my 7 year old knows you round up at the midpoint. Anyway, the app is 100% functional now with visual bugs and logic bugs fixed (that I am aware of) and I will likely deploy it.